新しいコンポーネントを追加し、注文ページのロジックを整理しました
